Moving to avro and simple packages
