+++ src/main/java/com/yahoo/bullet/record/avro/UntypedAvroBulletRecord.java
@@ -3,8 +3,10 @@
  *  Licensed under the terms of the Apache License, Version 2.0.
  *  See the LICENSE file associated with the project for terms.
  */
-package com.yahoo.bullet.record;
+package com.yahoo.bullet.record.avro;
 
+import com.yahoo.bullet.record.BulletRecord;
+import com.yahoo.bullet.record.UntypedBulletRecord;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
