+++ src/main/java/com/yahoo/bullet/record/simple/UntypedSimpleBulletRecordProvider.java
@@ -3,7 +3,10 @@
  *  Licensed under the terms of the Apache License, Version 2.0.
  *  See the LICENSE file associated with the project for terms.
  */
-package com.yahoo.bullet.record;
+package com.yahoo.bullet.record.simple;
+
+import com.yahoo.bullet.record.BulletRecord;
+import com.yahoo.bullet.record.BulletRecordProvider;
 
 /**
  * Factory for simple BulletRecords.
