+++ src/test/java/com/yahoo/bullet/record/simple/UntypedSimpleBulletRecordProviderTest.java
@@ -3,8 +3,9 @@
  *  Licensed under the terms of the Apache License, Version 2.0.
  *  See the LICENSE file associated with the project for terms.
  */
-package com.yahoo.bullet.record;
+package com.yahoo.bullet.record.simple;
 
+import com.yahoo.bullet.record.BulletRecord;
 import org.testng.Assert;
 import org.testng.annotations.Test;
